Make Allocator's RAII behavior more sensible.
We previously would check that there were no fallback blocks, which is equivalent to assuming that a user called FreeAll before the destructor was called. However, it's possible to get this wrong and things still work (for example, if no fallback blocks happen to be allocated since the latest call to FreeAll because the bump-ptr allocator was big enough).

Also, disallow copying, since that won't behave correctly (and I'm not sure it even make sense to copy a memory allocator -- that would violate pointer stability of the contents).

PiperOrigin-RevId: 247278977