Introduce a SidePair concept allowing us to rewrite much internal ruy code taking advantage of LHS<->RHS code symmetry to remove some redundancy.
The key motivation was that I want to experiment with some nontrivial
changes to how TrMulTask handles the packing of blocks, and I didn't want
to have to maintain two copies of this nontrivial code. With this change,
this code is now in a EnsurePacked method that's all I'll have to edit.

PiperOrigin-RevId: 259980220