Fix -Wsign-compare warnings.
Ruy uses int as the integer size / index type throughout.

It's enough because it's always at least 32 bit as far as ruy is concerned, and we don't cater to applications where matrices would exceed 2G.

It's nicer for us than to use std::size_t / std::ptrdiff_t because it's shorter ('int') so that makes code a little more readable, and secondarily there is a minor arithmetic efficiency benefit, especially in :size_util. If we supported 64bit, some code would have to be a little more complex, and then there would be a temptation to write a simpler faster 32bit variant, and now ruy would be more complicated, with more opportunities to accidentally cast 64bit integers to 32bit, etc.

There is one exception: in allocator.* we use std::ptrdiff_t because we have actually run into a use case (b/139735268) where the overall allocated size exceeded 2G even though no individual matrix did.

Prior to this CL, another exception was prepacked_cache.*. The code used size_t, however it did not do so consistently and had some harcoded eviction threshold at 2^28 bytes, safely below the limit for 32bit integers, so there did not seem to be a reason not to use int.

This CL changes prepacked_cache.* to use 'int' and makes other more localized changes in various places to avoid the warnings.

PiperOrigin-RevId: 305755191